refactor(VendingMachineProcessor) : 기능에 맞게 클래스 명 수정
