--- src/main/java/vendingmachine/service/MoneyChecker.java
@@ -1,7 +1,0 @@
-package vendingmachine.service;
-
-public enum MoneyChecker {
-
-    ENOUGH,
-    NOT_ENOUGH
-}
