+++ src/main/java/vendingmachine/service/VendingMachineChecker.java
@@ -1,0 +1,11 @@
+package vendingmachine.service;
+
+import vendingmachine.domain.Money;
+
+import java.util.function.Consumer;
+
+public enum VendingMachineChecker {
+
+    OPEN,
+    CLOSE
+}
